chore: renames controller to controllers
