+++ controllers/session/model_convert_test.go
@@ -2,7 +2,7 @@ package session_test
 
 import (
 	"github.com/maistra/istio-workspace/api/maistra/v1alpha1"
-	"github.com/maistra/istio-workspace/controller/session"
+	"github.com/maistra/istio-workspace/controllers/session"
 	"github.com/maistra/istio-workspace/pkg/model"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
