+++ controllers/session/session_controller_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/maistra/istio-workspace/api/maistra/v1alpha1"
-	"github.com/maistra/istio-workspace/controller/session"
+	"github.com/maistra/istio-workspace/controllers/session"
 	"github.com/maistra/istio-workspace/pkg/model"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
